requirements.txt
# TB Detection Delay Multi-Method Analysis
# Requirements for reproducing the analysis framework

# Core data science libraries
pandas>=2.1.3
numpy>=1.24.0
scipy>=1.10.0

# Visualization libraries
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.15.0

# Statistical and machine learning
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Probabilistic programming (Bayesian MCMC analysis)
# Note: PyMC not compatible with Python 3.14, using NumPyro instead
numpyro>=0.13.2
jax>=0.4.20
jaxlib>=0.4.20
arviz>=0.16.1

# Graph theory and network analysis (DAG modeling)
networkx>=3.1

# Document processing and reporting
python-docx>=1.1.0
openpyxl>=3.1.0

# Data access and web utilities
requests>=2.31.0
urllib3>=2.0.0

# Geospatial analysis
geopandas>=0.13.0
shapely>=2.0.0

# Configuration and data formats
pyyaml>=6.0
xlrd>=2.0.0

# Development and reproducibility
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For advanced users
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
